Fix signed/unsigned comparison warning in wxAuiMDIClientWindow
This was introduced in 73f8a6dacb7e675b79a2f3eb6f5212f1cc8a1e58, see
https://github.com/wxWidgets/wxWidgets/pull/802

(cherry picked from commit 67cb0ad26043270c1fb9b9681a4d365a9e2f4d57)

+++ src/aui/tabmdi.cpp
@@ -820,7 +820,7 @@ int wxAuiMDIClientWindow::SetSelection(size_t nPage)
 wxAuiMDIChildFrame* wxAuiMDIClientWindow::GetActiveChild()
 {
     const int sel = GetSelection();
-    if ( sel == wxNOT_FOUND || sel >= GetPageCount() )
+    if ( sel == wxNOT_FOUND || sel >= (int)GetPageCount() )
         return NULL;
 
     return wxStaticCast(GetPage(sel), wxAuiMDIChildFrame);
